+++ diffpy/snmf/componentsignal.py
